Move common type declarations to a new file
