+++ src/api/cache.ts
@@ -1,6 +1,6 @@
 import { createHash } from 'crypto';
 import Markpilot from 'src/main';
-import { APIClient, ChatMessage } from './openai';
+import { APIClient, ChatMessage } from './types';
 
 export class MemoryCache implements APIClient {
   private store: Map<string, string> = new Map();
