+++ src/chat/components/ChatItem.tsx
@@ -2,7 +2,7 @@ import { Bot, Copy, User } from 'lucide-react';
 import ReactMarkdown from 'react-markdown';
 import rehypeKatex from 'rehype-katex';
 import remarkMath from 'remark-math';
-import { ChatMessage } from 'src/api/openai';
+import { ChatMessage } from 'src/api/types';
 
 export function ChatItem({ message }: { message: ChatMessage }) {
   return (
